+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Update to latest Polkadot, Kusama & Substrate metadata
+
+
 ## 9.9.1 Nov 20, 2022
 
 Changes: